requirements.txt
aiofiles==0.8.0
aiohttp==3.7.4.post0
aiosignal==1.2.0
aiosqlite==0.17.0
Async-OAuthlib==0.0.9
async-timeout==3.0.1
attrs==21.4.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
colorlog==6.6.0
discord-ext-ipc==2.1.1
discord.py==1.7.3
frozenlist==1.3.0
h11==0.13.0
h2==4.1.0
hikari==2.0.0.dev106
hikari-lightbulb==2.2.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==6.0.2
oauthlib==3.2.0
priority==2.0.0
PyJWT==2.3.0
python-decouple==3.6
quart==0.16.3
Quart-Discord==2.1.4
requests==2.27.1
toml==0.10.2
typing-extensions==4.1.0
urllib3==1.26.8
Werkzeug==2.0.3
wsproto==1.0.0
yarl==1.7.2
